+++ mc2/controllers/docker/forms.py
@@ -17,6 +17,11 @@ class DockerControllerForm(ControllerForm):
             'class': 'form-control',
             'placeholder': '(optional)'}),
         required=False)
+    marathon_health_check_cmd = forms.CharField(
+        widget=forms.TextInput(attrs={
+            'class': 'form-control',
+            'placeholder': '(optional)'}),
+        required=False)
     port = forms.CharField(
         widget=forms.NumberInput(attrs={
             'class': 'form-control',
@@ -57,7 +62,8 @@ class DockerControllerForm(ControllerForm):
             'port', 'domain_urls', 'volume_needed', 'volume_path',
             'webhook_token', 'description', 'organization',
             'postgres_db_needed', 'external_visibility',
-            'rabbitmq_vhost_needed', 'rabbitmq_vhost_name')
+            'rabbitmq_vhost_needed', 'rabbitmq_vhost_name',
+            'marathon_health_check_cmd',)
 
 
 class DockerControllerFormHelper(ControllerFormHelper):